Moved old orionldGetEntities from serviceRoutines to legacyDriver
