+++ test/functionalTest/cases/0000_ngsild/ngsild_langprop-batch-create.test
@@ -80,7 +80,7 @@ echo
 
 echo "03. GET both entities"
 echo "====================="
-orionCurl --url /ngsi-ld/v1/entities?type=T
+orionCurl --url /ngsi-ld/v1/entities?type=T -H "Orionld-Legacy: true"
 echo
 echo
